Start coallesce
